requirements.txt
selenium==4.17.2
selenium-wire==5.1.0
seleniumbase==4.23.2
llama-cpp-python==0.1.78
huggingface_hub
